requirements.txt
python-dotenv
wikipedia
tiktoken
neo4j
streamlit
Pillow
fastapi
PyPDF2
pydantic
uvicorn
sse-starlette
boto3
streamlit==1.32.1
# missing from the langchain base image?
langchain-openai==0.2.4
langchain-community==0.3.3
langchain-google-genai==2.0.3
langchain-ollama==0.2.0
langchain-huggingface==0.1.1
langchain-aws==0.2.4
nltk
streamlit_modal
streamlit_autorefresh